backend/requirements.txt
fastapi
uvicorn
httpx
websockets
python-dotenv
pydantic
pydantic-settings
